requirements.txt
tgcrypto>=1.2.5
aiohttp>=3.8.0
pytelegrambotapi>=4.6.0
pyrotgfork>=2.3.20
python-binance>=1.0.19
google-genai>=0.7.0
aiofiles>=23.1.0
aiogram>=3.4.0
prettytable>=3.9.0
requests>=2.31.0
asyncio-mqtt>=0.16.1
urllib3>=1.26.18
typing-extensions>=4.8.0